pyproject.toml
[tool.poetry]
name = "asyncnotes"
version = "0.1.0"
description = ""
authors = ["Madiyar <madiarashkenovm@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.30.6"
sqlalchemy = "^2.0.35"
asyncpg = "^0.29.0"
alembic = "^1.13.2"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "^3.3.0"
python-multipart = "^0.0.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
